+++ BxInputController/Sources/Rows/Date/View/BxInputSelectorDateRowBinder.swift
@@ -14,7 +14,7 @@
 import UIKit
 
 /// Binder for BxInputSelectorDateRow subclasses (parent selector)
-open class BxInputSelectorDateRowBinder<Row: BxInputSelectorDateRow, Cell: BxInputSelectorCell>: BxInputSelectorRowBinder<Row, Cell>
+open class BxInputSelectorDateRowBinder<Row: BxInputSelectorDateRow, Cell: BxInputSelectorCell>: BxInputSelectorRowBinder<Row, Cell>, BxInputRowBinderMenuAll
 {
     
     /// visual updating of value && separator that depends on type of the row
@@ -25,6 +25,34 @@ open class BxInputSelectorDateRowBinder<Row: BxInputSelectorDateRow, Cell: BxInp
         } else {
             cell?.valueTextField.text = ""
         }
+    }
+    
+    open func copyValue() {
+        if let date = row.value {
+            UIPasteboard.general.string = owner?.settings.dateFormat.string(from: date)
+        } else {
+            UIPasteboard.general.string = ""
+        }
+    }
+    
+    open func pasteValue() {
+        if let string = UIPasteboard.general.string, let date = owner?.settings.dateFormat.date(from: string) {
+            row.value = date
+            update()
+            didChangeValue()
+            if row.isOpened {
+                if let child = owner?.getRowBinder(for: row.child) as? BxInputChildSelectorDateRowBinder {
+                    child.updateDate(animated: true)
+                }
+            }
+        }
+    }
+    
+    open func deleteValue() {
+        row.value = nil
+        update()
+        didChangeValue()
+        toClose()
     }
     
 }